+++ packages/components/src/components/action-row/action-row.styles.ts
@@ -69,6 +69,7 @@ export const actionRowStyles = css`
     display: inline-flex;
     align-items: center;
     justify-content: center;
+    color: var(--action-row-icon);
   }
 
   minis-checkbox {